refactor functions
